Moved JobIDToName routine back to CustomComboInfoAttribute Fixed Blue Mage Spellbook warning not showing to non-English clients due to recent jobname localization change
